refactor: move `const_names` macro to dedicated module
